requirements.txt

aws-cdk-lib==2.87.0
aws-lambda-powertools==2.20.0
boto3==1.28.5
botocore==1.31.5
constructs==10.2.69
freezegun==1.2.2
pydantic==2.0.3
pydantic_core==2.3.0
pytest==7.4.0
python-dateutil==2.8.2
slack-sdk==3.21.3